vec geom incorporated
