+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+## [2019.06]
+### Added
+- Vector geometry added to 3D case replacing the scalar aproach
+- Center of Rotation (CenterRotOffset variable) can be defined in the class as a scalar to avoid direct data manupulations (cropping, padding)
+
+### Changed
+- Demos for 2D/3D reconstruction updated with a new class
+
+### Removed
+- Section about "changelog" vs "CHANGELOG".
